# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SmartSchool Asset API",
        "description": "Dokumentasi API untuk sistem manajemen aset SmartSchool. Mencakup peminjaman barang, pengambilan konsumsi, barang masuk, dan laporan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/rooms": {
            "get": {
                "tags": [
                    "Ruangan"
                ],
                "summary": "Daftar semua ruangan (admin)",
                "description": "Menampilkan daftar semua ruangan yang terdaftar di sistem. Hanya untuk admin.",
                "operationId": "fc04f85434400dbf0e26d2ec77a5878d",
                "responses": {
                    "200": {
                        "description": "Halaman daftar ruangan"
                    },
                    "403": {
                        "description": "Akses ditolak"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ruangan"
                ],
                "summary": "Simpan ruangan baru (admin)",
                "description": "Menyimpan data ruangan baru ke sistem. Kode ruangan harus unik.",
                "operationId": "afb43684c849533c65e931df636b4de6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_ruangan",
                                    "nama_ruangan"
                                ],
                                "properties": {
                                    "kode_ruangan": {
                                        "description": "Kode unik ruangan",
                                        "type": "string",
                                        "example": "R-01"
                                    },
                                    "nama_ruangan": {
                                        "type": "string",
                                        "example": "Ruang Guru"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect ke daftar ruangan jika berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal (kode sudah dipakai)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rooms/create": {
            "get": {
                "tags": [
                    "Ruangan"
                ],
                "summary": "Form tambah ruangan (admin)",
                "description": "Menampilkan form untuk menambahkan ruangan baru. Hanya untuk admin.",
                "operationId": "56a94108c99bd608ba2873380a6f9ec5",
                "responses": {
                    "200": {
                        "description": "Halaman form tambah ruangan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rooms/{id}/edit": {
            "get": {
                "tags": [
                    "Ruangan"
                ],
                "summary": "Form edit ruangan (admin)",
                "description": "Menampilkan form untuk mengedit data ruangan. Hanya untuk admin.",
                "operationId": "5ee8ee2dd7ee1a6437d0434e32d6ae9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Halaman form edit ruangan"
                    },
                    "404": {
                        "description": "Ruangan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rooms/{id}": {
            "put": {
                "tags": [
                    "Ruangan"
                ],
                "summary": "Update data ruangan (admin)",
                "description": "Mengupdate data ruangan. Kode ruangan harus unik kecuali milik ruangan itu sendiri.",
                "operationId": "ea75b11760a871bba18cf5f9e43a5b70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kode_ruangan",
                                    "nama_ruangan"
                                ],
                                "properties": {
                                    "kode_ruangan": {
                                        "type": "string",
                                        "example": "R-01"
                                    },
                                    "nama_ruangan": {
                                        "type": "string",
                                        "example": "Ruang Kepala Sekolah"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Redirect ke daftar ruangan jika berhasil"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "404": {
                        "description": "Ruangan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Ruangan"
                ],
                "summary": "Hapus ruangan (admin)",
                "description": "Menghapus data ruangan dari sistem. Hanya untuk admin.",
                "operationId": "cf9104dbefd91a81bebb07cafe7e70cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect dengan pesan sukses"
                    },
                    "404": {
                        "description": "Ruangan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token autentikasi Laravel Sanctum/Session",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}